procutil: don't send signals to nil process (#2826)

+++ pkg/procutil/procutil_unix.go
@@ -13,12 +13,18 @@ func SetOptNewProcessGroup(attrs *syscall.SysProcAttr) {
 }
 
 func KillProcessGroup(cmd *exec.Cmd) {
-	if cmd != nil && cmd.Process != nil {
-		// Kill the entire process group.
-		_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
+	if cmd == nil || cmd.Process == nil {
+		return
 	}
+
+	// Kill the entire process group.
+	_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
 }
 
 func GracefullyShutdownProcess(p *os.Process) error {
+	if p == nil {
+		return nil
+	}
+
 	return syscall.Kill(-p.Pid, syscall.SIGTERM)
 }
